+++ src/control_plane/shared_infrastructure/gotenberg/mod.rs
@@ -1,0 +1,11 @@
+pub(crate) use gotenberg_plan_error::GotenbergPlanError;
+pub(crate) use gotenberg_project_resources::GotenbergProjectResources;
+pub(crate) use gotenberg_shared_instance_plan::GotenbergSharedInstancePlan;
+pub(crate) use gotenberg_shared_instance_plan_options::GotenbergSharedInstancePlanOptions;
+pub(crate) use plan_gotenberg_project_resources::plan_gotenberg_project_resources;
+
+mod gotenberg_plan_error;
+mod gotenberg_project_resources;
+mod gotenberg_shared_instance_plan;
+mod gotenberg_shared_instance_plan_options;
+mod plan_gotenberg_project_resources;